[MODFIN-189] - Move tests to appropriate packages
