+++ src/test/java/org/folio/ApiTestSuite.java
@@ -18,7 +18,14 @@ import org.folio.services.fiscalyear.FiscalYearServiceTest;
 import org.folio.services.fund.FundDetailsServiceTest;
 import org.folio.services.budget.CreateBudgetServiceTest;
 import org.folio.services.fund.FundServiceTest;
+import org.folio.services.group.GroupExpenseClassTotalsServiceTest;
 import org.folio.services.group.GroupFundFiscalYearServiceTest;
+import org.folio.services.ledger.LedgerDetailsServiceTest;
+import org.folio.services.ledger.LedgerRolloverErrorsServiceTest;
+import org.folio.services.ledger.LedgerRolloverProgressServiceTest;
+import org.folio.services.ledger.LedgerRolloverServiceTest;
+import org.folio.services.ledger.LedgerServiceTest;
+import org.folio.services.ledger.LedgerTotalsServiceTest;
 import org.folio.services.protection.AcqUnitMembershipsServiceTest;
 import org.folio.services.protection.AcqUnitsServiceTest;
 import org.folio.services.protection.ProtectionServiceTest;
