+++ src/test/java/org/folio/rest/impl/LedgerRolloverApiTest.java
@@ -29,7 +29,6 @@ import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.LedgerFiscalYearRollover;
 import org.folio.rest.jaxrs.model.LedgerFiscalYearRolloverCollection;
 import org.folio.rest.util.TestEntities;
-import org.folio.services.ledger.LedgerRolloverService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
