+++ src/test/java/org/folio/rest/impl/LedgersApiTest.java
@@ -47,8 +47,6 @@ import org.folio.rest.jaxrs.model.Errors;
 import org.folio.rest.jaxrs.model.Ledger;
 import org.folio.rest.jaxrs.model.LedgersCollection;
 import org.folio.rest.util.TestEntities;
-import org.folio.services.ledger.LedgerDetailsService;
-import org.folio.services.ledger.LedgerService;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
