+++ src/test/java/org/folio/services/ledger/LedgerRolloverErrorsServiceTest.java
@@ -1,10 +1,9 @@
-package org.folio.services;
+package org.folio.services.ledger;
 
 import org.folio.rest.acq.model.finance.LedgerFiscalYearRolloverErrorCollection;
 import org.folio.rest.core.RestClient;
 import org.folio.rest.core.models.RequestContext;
 import org.folio.rest.exception.HttpException;
-import org.folio.services.ledger.LedgerRolloverErrorsService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.platform.runner.JUnitPlatform;
